main/requirements.txt
absl-py==0.10.0
appnope==0.1.0
argon2-cffi==20.1.0
astropy==4.0.1.post1
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
confuse==1.3.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fitparse==1.1.0
gast==0.3.3
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
htmlmin==0.1.12
html2text==2020.1.16
idna==2.10
ImageHash==4.1.0
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
llvmlite==0.34.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
missingno==0.4.2
mistune==0.8.4
mlxtend==0.17.3
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.1.3
numba==0.51.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.0
pandas-profiling==2.8.0
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
phik==0.10.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
pymysql==1.0.2
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
spicy==0.16.0
tangled-up-in-unicode==0.0.6
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
torch===1.8.1
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
tzwhere==3.0.3
urllib3==1.25.10
visions==0.4.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.1.1
zipp==3.1.0
torch===1.8.1